[feat/#95] 헤더 딜레이 수정
